test(analyzers): red tests for antislop matcher bugs
Batch HB W3: move #434/#435/#438 contracts from strict xfails in
test_cov_antislop_matcher_paths into test_antislop_matcher_hb with
non-strict W4 markers per D2.
