+++ tests/analyzers/test_cov_antislop_matcher_paths.py
@@ -465,103 +465,3 @@ def test_snippet_falls_back_to_the_truncated_text_when_the_pattern_misses() -> N
     text = "x" * 400
 
     assert _snippet(text, re.compile("never-here")) == "x" * 120
-
-
-# --- known defects -------------------------------------------------------------
-
-
-@pytest.mark.xfail(
-    reason=(
-        "_call_positional_argument_names silently drops positional arguments that are "
-        'not identifiers or attributes, so `return request("GET", url)` is read as '
-        "forwarding exactly `(url,)` and the wrapper is reported as adding nothing. "
-        "Binding a constant argument is real behaviour; this fires on ordinary adapter "
-        "functions. Keyword arguments already abort the check (`return ()`) — literal "
-        "positionals must too."
-    ),
-    strict=True,
-)
-def test_wrapper_that_binds_a_literal_argument_is_not_a_pass_through() -> None:
-    """A wrapper supplying an extra constant argument adds behaviour."""
-    source = "def fetch(url: str) -> object:\n    return request('GET', url)\n"
-
-    assert "antislop/pass-through-wrapper" not in _rule_ids(source)
-
-
-@pytest.mark.xfail(
-    reason=(
-        "_node_text slices the source string with tree-sitter BYTE offsets, so any "
-        "non-ASCII character earlier in the file shifts every later node's text. Here "
-        "the binding name for `import os` is read as 's\\n', which no body reference "
-        "can match, and a used import is reported as phantom. mergeCraft's own sources "
-        "are full of em dashes, so this fires on ordinary files."
-    ),
-    strict=True,
-)
-def test_non_ascii_above_an_import_must_not_make_a_used_import_phantom() -> None:
-    """A used import stays used when the file contains a non-ASCII character."""
-    source = "# café — notes\nimport os\n\n\ndef run() -> str:\n    return os.sep\n"
-
-    assert "antislop/phantom-import" not in _rule_ids(source)
-
-
-@pytest.mark.xfail(
-    reason=(
-        "_node_text byte/char mismatch again: the reported snippet is a shifted slice "
-        "of the source ('ise NotImplementedError'), so the review comment quotes text "
-        "that does not exist in the file."
-    ),
-    strict=True,
-)
-def test_snippet_after_non_ascii_quotes_real_source_text() -> None:
-    """The placeholder snippet must be a substring of the file it came from."""
-    source = "# — notes\ndef pending(value: int) -> int:\n    raise NotImplementedError\n"
-
-    findings = _findings(source)
-    assert findings == [("antislop/placeholder-implementation", 2, "raise NotImplementedError")]
-
-
-@pytest.mark.xfail(
-    reason=(
-        "antislop/empty-error-handler never fires on Python. tree-sitter-python's "
-        "except_clause exposes its suite as an unnamed child, so "
-        "child_by_field_name('body') is always None and the whole matcher body is "
-        "dead. The rule declares languages [python, javascript, typescript] and ships "
-        "a positive Python fixture; test_positive_fixture_fires_rule passes only "
-        "because the .ts fixture matches."
-    ),
-    strict=True,
-)
-def test_python_except_block_that_only_passes_is_reported() -> None:
-    """``except OSError: pass`` swallows the failure and must be reported."""
-    source = (
-        "def load() -> str:\n"
-        "    try:\n"
-        "        return open('data.txt', encoding='utf-8').read()\n"
-        "    except OSError:\n"
-        "        pass\n"
-        "    return 'fallback'\n"
-    )
-
-    assert "antislop/empty-error-handler" in _rule_ids(source)
-
-
-@pytest.mark.xfail(
-    reason=(
-        "antislop/error-obscuring-catch never fires on Python — same dead "
-        "child_by_field_name('body') lookup on except_clause as empty-error-handler. "
-        "Its positive Python fixture is likewise carried by the .ts sibling."
-    ),
-    strict=True,
-)
-def test_python_except_block_returning_none_is_reported() -> None:
-    """``except KeyError: return None`` hides the failure and must be reported."""
-    source = (
-        "def lookup(key: str) -> str | None:\n"
-        "    try:\n"
-        "        return _fetch(key)\n"
-        "    except KeyError:\n"
-        "        return None\n"
-    )
-
-    assert "antislop/error-obscuring-catch" in _rule_ids(source)
